refactor(ui): replace RecordingControls with RecordingDeviceSelector and TransformationSelector for improved UI structure
