+++ apps/app/src/lib/components/recording-controls/_combobox.ts
@@ -2,5 +2,3 @@ import { useCombobox } from '../useCombobox.svelte';
 
 // Create and export the combobox instance here to avoid circular dependencies
 export const combobox = useCombobox();
-
-export { default as RecordingControls } from './RecordingControls.svelte';
